chore: Use `vscode.Uri` everywhere with better path math
TIL that for supporting remote file URIs, it's best to always use the URI
and to do path math with methods on the URI object or via
`vscode.Uri.joinPath`. This keeps the original URI in-tact, e.g.
`vscode.Uri.joinPath(uriDir, "..")` to get the parent `uriDir`.
Another useful approach is `uri.with({ path: newPath })` to
modify only the path component of the uri, e.g.
`uri.with({ path: path.dirname(uri.path) })` to get parent folder.